python/requirements.txt
google-genai>=1.0.0
pyaudio>=0.2.14
numpy>=1.24.0
python-dotenv>=1.0.0
